#71 support user-defined instance and static paths
This is necessary because otherwise, the files uploaded during testing are stored in the same folder as during production. The current config is still not great, and I'm also not sure whether static files are working 100% correctly when the static path is changed. However, so far it has worked during testing of file upload/download.