#104 #105 greatly simplify site banner
The site banner is now a simple, clickable image which uses `<picture>` with two sources in order to responsively choose which image to choose (featured vs. banner). This means we were able to get rid of the script.

Unfortunately I have not been able to overlay the "Stefan on Software" text.